remove args from deserializationlistener; more doc on utils..
